+++ packages/hippo_core/hippo_core_flutter/lib/hippo_core_flutter.dart
@@ -14,5 +14,6 @@ export 'src/data_subject/data_subject.dart';
 export 'src/filter.dart';
 export 'src/layout.dart';
 export 'src/resource.dart';
+export 'src/runtime_platform.dart';
 export 'src/sliver.dart';
 export 'src/store/store.dart';